featured product, campaign product card and campaign timer finished

+++ lib/view/utils/constant_styles.dart
@@ -211,7 +211,7 @@ Widget seeAllTitle(String title) {
   );
 }
 
-Widget discAmountRow(double discountAmount, double amount) {
+Widget discAmountRow(int discountAmount, int amount) {
   return Row(
     children: [
       Text(
